requirements_all.txt
numpy==1.26.4
pandas==2.2.3
pickle4==0.0.1
flask==3.0.3
python-dotenv==1.0.1
opencv-python==4.10.0.84
face-recognition==1.3.0
Werkzeug==3.0.3     # Make sure this is exact same version, this is the trouble maker!!!!!!!
openpyxl==3.1.5
pprintpp            # Optional
